backend/requirements.txt
azure-ai-projects>=1.0.0b11
azure-identity>=1.15.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0